+++ src/avx/avx64_butterflies.rs
@@ -4,7 +4,7 @@ use std::mem::MaybeUninit;
 
 use num_complex::Complex;
 
-use crate::common::FFTnum;
+use crate::common::FftNum;
 
 use crate::{Direction, Fft, FftDirection, Length};
 
@@ -34,7 +34,7 @@ macro_rules! boilerplate_fft_simd_butterfly {
                 }
             }
         }
-        impl<T: FFTnum> Fft<T> for $struct_name<f64> {
+        impl<T: FftNum> Fft<T> for $struct_name<f64> {
             fn process_with_scratch(
                 &self,
                 input: &mut [Complex<T>],
@@ -212,7 +212,7 @@ macro_rules! boilerplate_fft_simd_butterfly_with_scratch {
                 unsafe { self.row_butterflies(RawSlice::new(output), RawSliceMut::new(output)) };
             }
         }
-        impl<T: FFTnum> Fft<T> for $struct_name<f64> {
+        impl<T: FftNum> Fft<T> for $struct_name<f64> {
             fn process_with_scratch(
                 &self,
                 input: &mut [Complex<T>],